New Feature: CollectionVersions
We found a need to allow multiple versions of the same collection. In this commit we introduce a new model 'CollectionVersion'. We've also added some new attributes for ease of use in the Django admin. As with any new feature, this may not be the only commit related to CollectionVersions as the new feature is implemented. Any current use of dataforms should not be affected as this is an optional feature. 

Basic use case: We have an application on our website which uses the  'gen-application' collection. People fill out this application. Then we have a new updated version of 'gen-application', but the old 'gen-application' needs to remain intact for reference and historical purposes. So, we would create the new 'gen-applicaiton-2' as a separate collection from the original. Then we add an entry to CollectionVersion with a slug of 'gen-application' but point the Collection FK in that model to the new 'gen-application-2'. When we need to create a new application, we merely consult CollectionVersion for the current/newest version then use the referenced collection for the new application. This way all old applications will still point to the old collection in their submission objects while any new ones will end up referencing the new collection in their submission objects. 

Side note: If the old collection needs to stay the same, then all forms, fields & bindings most likely need to stay the same as well. So, if any field changes, you need to create a new field and a new form that uses the new version of the field. (You can always reference the old fields or forms if they have not changed.) Also remember to create any necessary bindings along the way. Then the new collection would use the new form. (If this doesn't make sense now, hopefully it will if/when you need this functionality.)


git-svn-id: http://django-dataforms.googlecode.com/svn/trunk@146 23ed9a0e-3f40-11de-b5fc-91815ed80701